+++ src/cds_resource.cc
@@ -146,7 +146,7 @@ std::string CdsResource::encode()
 std::shared_ptr<CdsResource> CdsResource::decode(const std::string& serial)
 {
     std::vector<std::string> parts = splitString(serial, RESOURCE_PART_SEP, true);
-    int size = parts.size();
+    size_t size = parts.size();
     if (size < 2 || size > 4)
         throw_std_runtime_error("Could not parse resources");
